refactor(driver): extract PendingL1SubmissionQueue to driver/pending_queue.rs (refactor 2.1c)
Third commit of the Phase 2.1 mechanical split. Moves the
PendingL1SubmissionQueue (Phase 1.5's artifact) and its closely-related
types out of driver/mod.rs into a dedicated sibling module.

- `crates/based-rollup/src/driver/pending_queue.rs` (new, 202 lines):
  - `TriggerMetadata` — L1 trigger group metadata for `executeL2TX`.
  - `PendingL1Group` — a single trigger group.
  - `BlockEntryMix` — the `Empty/OnlyDeposits/OnlyWithdrawals/Mixed`
    classifier (closes invariant #11 at the type level).
  - `PendingL1SubmissionQueue` + impl — the 4-in-1 unified queue
    (append_group, truncate_to, clear, trim_entries_from_front,
    entry_mix).
  - Module comment lifted verbatim from the pre-extraction section.
  - Imports: `CrossChainExecutionEntry`, `Address`, `U256`.

- `crates/based-rollup/src/driver/mod.rs` (-190 lines → 5127 total):
  - `mod pending_queue;` declaration.
  - `pub use pending_queue::{BlockEntryMix, PendingL1Group,
    PendingL1SubmissionQueue, TriggerMetadata};` — the re-export
    preserves `use super::*` in `driver_tests.rs`.
  - `Address` removed from the top-level `alloy_primitives` use — no
    other reference to `Address` remained in mod.rs after the move.

- `cargo check`, `cargo clippy --all-features -- -D warnings`, and
  `cargo nextest -p based-rollup` (608/608) all pass. Zero behavioral
  change; the `FlushPlan::submit_via` site still borrows `pending_l1`
  via the re-exported type path.

Next: 2.1d extracts `FlushPlan<S>` + `SendResult` + `RollbackPackage`
+ the sealed `Sendable` trait to `driver/flush_plan.rs`.
